Added Anglerfish, Edited Inventory Manager
Added the Anglerfish sprite as a temporary image.
Added the logic that after reaching a certain range, the fish is attracting the player. The pull effect is consistent.
Also changed the EnemySpawner Script so that AnglerFish spawns instead of Pufferfish.
In addition added Some new Formatings to the InventoryManager Script in the Inspector View.