First Concept of GameOverScreen
*** Firstly Changed the Name INVENTORYCANVAS to GAMECANVAS.
Added the GameOverScreen as an Background within the GameCanvas
Added a GameOverScreen which will be triggered as soon as the Player dies. (For now the Game is on death paused, if we will have some fancy animations later on on the death screen, we have to adjust this mechanic)
Respawn is (for now) at the default start/spawn of the Player.
Will be later fixed so Player spawns at checkpoints after reaching a certain progression.